+++ pkg/market/client_test.go
@@ -7,16 +7,21 @@ import (
 	"testing"
 
 	"github.com/joho/godotenv"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/KazanExpress/yandex-market/pkg/market/models"
 )
 
 func TestMain(t *testing.M) {
 	_ = godotenv.Load()
-	t.Run()
+
+	os.Exit(t.Run())
 }
 
 func getOptions() Options {
-	var token = os.Getenv("OAUTH_TOKEN")
-	var clientID = os.Getenv("OAUTH_CLIENT_ID")
+	token := os.Getenv("OAUTH_TOKEN")
+	clientID := os.Getenv("OAUTH_CLIENT_ID")
+
 	return Options{
 		OAuthClientID: clientID,
 		OAuthToken:    token,
@@ -29,14 +34,16 @@ func getClient() *YandexMarketClient {
 }
 
 func getCampaign() int64 {
-	var camp = os.Getenv("CAMPAIGN")
+	camp := os.Getenv("CAMPAIGN")
 	if camp == "" {
 		return 1
 	}
-	var res, err = strconv.ParseInt(camp, 10, 64)
+
+	res, err := strconv.ParseInt(camp, 10, 64)
 	if err != nil {
 		return 2
 	}
+
 	return res
 }
 
@@ -44,9 +51,11 @@ func TestYandexMarketClient_ListFeeds(t *testing.T) {
 	type fields struct {
 		options Options
 	}
+
 	type args struct {
 		campaignID int64
 	}
+
 	tests := []struct {
 		name    string
 		args    args
@@ -60,6 +69,7 @@ func TestYandexMarketClient_ListFeeds(t *testing.T) {
 			wantLen: 5,
 		},
 	}
+
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			c := getClient()
@@ -80,15 +90,17 @@ func TestYandexMarketClient_RefreshFeed(t *testing.T) {
 		campaignID int64
 		feedID     int64
 	}
+
 	type test struct {
 		name    string
 		args    args
 		wantErr bool
 	}
-	var tests = []test{}
-	var c = getClient()
-	var campaignID = getCampaign()
-	var feeds, err = c.ListFeeds(campaignID)
+
+	tests := []test{}
+	c := getClient()
+	campaignID := getCampaign()
+	feeds, err := c.ListFeeds(campaignID)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -108,4 +120,86 @@ func TestYandexMarketClient_RefreshFeed(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestYandexMarketClient_Prices(t *testing.T) {
+	c := getClient()
+	campaignID := getCampaign()
+	offerID := "169690W424150"
+	feedID := int64(820450)
+	discountBase := 300.0
+	price := 250.0
+
+	err := c.SetOfferPrices(campaignID, []models.Offer{
+		{
+			Feed:   models.FeedObj{ID: feedID},
+			Delete: false,
+			ID:     offerID,
+			Price: models.Price{
+				CurrencyID:   "RUR",
+				DiscountBase: discountBase,
+				Value:        price,
+			},
+		},
+	})
+
+	assert.NoError(t, err)
+
+	offerPrices, err := c.GetOfferPrices(campaignID, nil, nil)
+
+	assert.NoError(t, err)
+	assert.Len(t, offerPrices, 1, "there should be only 1 product set")
+
+	offerPrice := offerPrices[0]
+
+	assert.Equal(t, offerID, offerPrice.ID, "ids should match")
+	assert.Equal(t, discountBase, offerPrice.Price.DiscountBase, "discountBase should match")
+	assert.Equal(t, price, offerPrice.Price.Value, "price should match")
+
+	// WARN: uncomment accurately
+	// err = c.DeleteAllOffersPrices(campaignID)
+	// assert.NoError(t, err)
+
+	// offerPrices, err = c.GetOfferPrices(campaignID, nil, nil)
+
+	// assert.NoError(t, err)
+	// assert.Len(t, offerPrices, 0, "there should be no product price set")
+}
+
+func TestYandexMarketClient_Hidden(t *testing.T) {
+	c := getClient()
+	campaignID := getCampaign()
+	offerID := "169690W424150"
+	feedID := int64(820450)
+	comment := "Временно закончился на складе"
+
+	err := c.HideOffers(campaignID, []models.HiddenOffer{
+		{
+			FeedID:     feedID,
+			OfferID:    offerID,
+			TTLInHours: 12,
+			Comment:    comment,
+		},
+	})
+	assert.NoError(t, err)
+
+	res, err := c.GetHiddenOffers(campaignID, "", nil, nil)
+
+	assert.NoError(t, err)
+	assert.NotZero(t, res.Total)
+
+	err = c.UnhideOffers(campaignID, []models.OfferToUnhide{
+		{
+			FeedID:  feedID,
+			OfferID: offerID,
+		},
+	})
+
+	assert.NoError(t, err)
+	assert.NotZero(t, res.Total)
+
+	res, err = c.GetHiddenOffers(campaignID, "", nil, nil)
+
+	assert.NoError(t, err)
+	assert.Zero(t, res.Total)
 }